Create README.md
This is a javascript based project which can be used to decide who is the winner in case of any situation. Refresh the website and enter the name of players accordingly. Then the random number generator generates random number for both the player and accordingly shows the image of the dice. And declares who won the game or is it a draw between both the players.